+++ src/__init__.py
@@ -1,4 +1,4 @@
 """Fexobooth - Photobooth Software für fexobox"""
 
-__version__ = "2.2.6"
+__version__ = "2.2.7"
 __author__ = "fexon e.K."